+++ apps/web/pages/flow/edges.tsx
@@ -1,0 +1,8 @@
+import { Edge } from "react-flow-renderer"
+
+const initialEdges: Edge[] = [
+  { id: 'e1-2', source: '1', target: '2' },
+  { id: 'e2-3', source: '2', target: '3', animated: true },
+]
+
+export default initialEdges
